requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.7.4
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.7
kombu==5.3.7
msgpack==1.0.8
mypy==1.11.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
types-beautifulsoup4==4.12.0.20240511
types-html5lib==1.1.11.20240228
types-PyYAML==6.0.12.20240724
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
